// hiauth-mgr-fornt/src/pages/sys/app-mgr/data.d.ts
export type AppListItem = {
  id: string;
  clientId: string;
  clientSecret: string;
  clientSecretExpiresAt: number;
  clientName: string;
  clientAuthenticationMethods: string;
  authorizationGrantTypes: string;
  redirectUris: string;
  scopes: string;
};

// export type AppListItem = {
//   id: number;
//   username: string;
//   phoneNum: string;
//   password: string;
//   avatar: string;
//   regtime: number;
//   status: number;
// };
//
// export type App = {
//   id: number;
//   username: string;
//   phoneNum: string;
//   avatar: string;
//   regtime: number;
//   status: number;
// };


// export type App = {
//   id: string;
//   clientId: string;
//   clientSecret: string;
//   clientSecretExpiresAt: number;
//   clientName: string;
//   clientAuthenticationMethods: string;
//   authorizationGrantTypes: string;
//   redirectUris: string;
//   scopes: string;
// }